The MultiDecoder project
This project is related to the generation of a multi-decoder network
merging the intra and the parallel MPEG-4 SP decoders. The project
contains the txt configuration file of the multi-decoder network and
also different cal files of the switching modules: the dummy cal files
(Sbox1x2 and Sbox2x1) instantiated in the merged network and examples of
static and dynamic implementations of these modules.